madrid-sample-workspace/client-extensions/liferay-sample-etc-python/requirements.txt
flask==2.1.3
requests
configparser
pyjwt
